+++ src/PostProcessingPasses.h
@@ -51,7 +51,7 @@ public:
 		CJitterHelper(JitterMode);
 		~CJitterHelper() {}
 
-		nm::float4x4 ComputeJitter(const uint64_t p_frameCount);
+		nm::float4x4 ComputeJitter(const uint32_t p_frameCount);
 		nm::float2 GetJitterOffset() { return m_jitterOffset; }
 
 	private:
@@ -61,7 +61,7 @@ public:
 
 		// Returns a single 2D point in a Hammersley sequence of length "numSamples", using base 1 and base 2
 		// Refer - https://mathworld.wolfram.com/HammersleyPointSet.html
-		inline nm::float2 Hammersley2D(uint64_t p_sampleIdx, uint64_t p_numSamples);
+		inline nm::float2 Hammersley2D(uint32_t p_sampleIdx, uint32_t p_numSamples);
 
 		// Computes a radical inverse with base 2 using crazy bit-twiddling from "Hacker's Delight"
 		// Refer - https://graphics.stanford.edu/~seander/bithacks.html
